server/requirements.txt
djangorestframework_simplejwt==4.6.0
~jango==3.1.12
Django==4.0.3
djangorestframework==3.13.1
